+++ src/onegov.gazette/onegov/gazette/collections/__init__.py
@@ -1,6 +1,8 @@
+from onegov.gazette.collections.categories import CategoryCollection
 from onegov.gazette.collections.notices import GazetteNoticeCollection
 
 
 __all__ = [
+    'CategoryCollection',
     'GazetteNoticeCollection'
 ]