+++ src/onegov.gazette/onegov/gazette/forms/__init__.py
@@ -1,0 +1,14 @@
+from onegov.gazette.forms.category import CategoryForm
+from onegov.gazette.forms.common import EmptyForm
+from onegov.gazette.forms.common import RejectForm
+from onegov.gazette.forms.notice import NoticeForm
+from onegov.gazette.forms.user import UserForm
+
+
+__all__ = [
+    'CategoryForm',
+    'EmptyForm',
+    'NoticeForm',
+    'RejectForm',
+    'UserForm',
+]