+++ src/onegov.gazette/onegov/gazette/forms/common.py
@@ -1,0 +1,19 @@
+from onegov.form import Form
+from onegov.gazette import _
+from wtforms import StringField
+from wtforms.validators import InputRequired
+
+
+class EmptyForm(Form):
+
+    pass
+
+
+class RejectForm(Form):
+
+    comment = StringField(
+        label=_("Comment"),
+        validators=[
+            InputRequired()
+        ]
+    )