+++ src/onegov.gazette/onegov/gazette/models/__init__.py
@@ -2,11 +2,13 @@ from onegov.gazette.models.notice import GazetteNotice
 from onegov.gazette.models.principal import Issue
 from onegov.gazette.models.principal import IssueDates
 from onegov.gazette.models.principal import Principal
+from onegov.gazette.models.category import Category
 
 
 __all__ = [
+    'Category',
     'GazetteNotice',
     'Issue',
     'IssueDates',
-    'Principal'
+    'Principal',
 ]